backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pydantic
websockets
